requirements.txt
openai-whisper
ffmpeg-python
google-generativeai
tqdm